refactor(frontend): decode where the kit instructions already live
Decoding inside the effect mapper put the token, token-2022 and system
packages into the chunk the activity list loads, since that mapper is
what the activity uses to build its own instruction list. 120KB for code
only the WalletConnect review ever runs, over the 100KB the size check
allows.

Moved to sol-instructions.utils, which is WalletConnect-only and already
holds every parser this needs. The three callers there decode their
message before handing it over, and the mapper goes back to
understanding one shape: the one the RPC reports.

That also drops the duplicate dispatch. Choosing a parser by program
address was written twice, once here and once as `parseSolInstruction`
in the module it moved to; the vocabulary now sits on top of that one
rather than beside a copy of it.

The shared mapper gains a single type import, which the build erases.
Measured with the check's own method: 14032KB either way.
