+++ src/frontend/src/sol/services/sol-simulation.services.ts
@@ -9,6 +9,7 @@ import type { SolanaNetworkType } from '$sol/types/network';
 import type { SolSimulationResult } from '$sol/types/sol-simulation';
 import type { CompilableTransactionMessage } from '$sol/types/sol-transaction-message';
 import { mapSolInstructionSummaries } from '$sol/utils/sol-instruction-summary.utils';
+import { asSolParsedRpcInstructionOrSelf } from '$sol/utils/sol-instructions.utils';
 import { deriveSolMessageSummary } from '$sol/utils/sol-message-summary.utils';
 import {
 	isEmptySolSimulationPreview,
@@ -92,7 +93,7 @@ const simulate = async ({
 	// The kit instructions are not parsed, so they contribute nothing themselves; iterating them is
 	// what attaches each simulated nested call to the instruction that made it.
 	const instructions = mapSolInstructionSummaries({
-		instructions: [...transactionMessage.instructions],
+		instructions: [...transactionMessage.instructions].map(asSolParsedRpcInstructionOrSelf),
 		innerInstructions: [...innerInstructions].map(({ index, instructions: inner }) => ({
 			index: Number(index),
 			instructions: [...inner]
@@ -110,7 +111,7 @@ const simulate = async ({
 	// the same transaction, which is what lets the review notice the two disagreeing.
 	const messageSummary = deriveSolMessageSummary({
 		instructions: mapSolInstructionSummaries({
-			instructions: [...transactionMessage.instructions],
+			instructions: [...transactionMessage.instructions].map(asSolParsedRpcInstructionOrSelf),
 			innerInstructions: [],
 			ownedAddresses: [address, ...ownedAddresses],
 			addressToToken
